fix codecheck

+++ device/services/profiler_service/src/trace_file_writer.cpp
@@ -48,20 +48,26 @@ bool TraceFileWriter::SetPluginConfig(const void* data, size_t size)
 
 void TraceFileWriter::SetTimeStamp()
 {
-    constexpr uint32_t NANOSECONDS = 1000000000;
+    constexpr uint64_t NANOSECONDS = 1000000000;
     struct timespec ts;
     clock_gettime(CLOCK_BOOTTIME, &ts);
-    header_.data_.boottime_ = ts.tv_sec * NANOSECONDS + ts.tv_nsec;
+    header_.data_.boottime_ = static_cast<uint64_t>(ts.tv_sec) * NANOSECONDS +
+        static_cast<uint64_t>(ts.tv_nsec);
     clock_gettime(CLOCK_REALTIME, &ts);
-    header_.data_.realtime_ = ts.tv_sec * NANOSECONDS + ts.tv_nsec;
+    header_.data_.realtime_ = static_cast<uint64_t>(ts.tv_sec) * NANOSECONDS +
+        static_cast<uint64_t>(ts.tv_nsec);
     clock_gettime(CLOCK_REALTIME_COARSE, &ts);
-    header_.data_.realtime_coarse_ = ts.tv_sec * NANOSECONDS + ts.tv_nsec;
+    header_.data_.realtime_coarse_ = static_cast<uint64_t>(ts.tv_sec) * NANOSECONDS +
+        static_cast<uint64_t>(ts.tv_nsec);
     clock_gettime(CLOCK_MONOTONIC, &ts);
-    header_.data_.monotonic_ = ts.tv_sec * NANOSECONDS + ts.tv_nsec;
+    header_.data_.monotonic_ = static_cast<uint64_t>(ts.tv_sec) * NANOSECONDS +
+        static_cast<uint64_t>(ts.tv_nsec);
     clock_gettime(CLOCK_MONOTONIC_COARSE, &ts);
-    header_.data_.monotonic_coarse_ = ts.tv_sec * NANOSECONDS + ts.tv_nsec;
+    header_.data_.monotonic_coarse_ = static_cast<uint64_t>(ts.tv_sec) * NANOSECONDS +
+        static_cast<uint64_t>(ts.tv_nsec);
     clock_gettime(CLOCK_MONOTONIC_RAW, &ts);
-    header_.data_.monotonic_raw_ = ts.tv_sec * NANOSECONDS + ts.tv_nsec;
+    header_.data_.monotonic_raw_ = static_cast<uint64_t>(ts.tv_sec) * NANOSECONDS +
+        static_cast<uint64_t>(ts.tv_nsec);
 }
 
 bool TraceFileWriter::Open(const std::string& path)
